+++ README.md
@@ -1,14 +1,38 @@
-<<h1>Pablos First Spacegame</h1>
+# Welcome to Pablo's First Spacegame 🦇
+
 This is Pablo's first awesome Spacegame.
-<h2>Requerements</h2>
-<b>Python:</b>
-To run it you neet to have Python installed.
+
+## Installation 🛠️
+
+First you need to install Python:
+
+```sh
 Go to https://www.python.org/downloads/ and download the latest version for your operating system
-<b>Python Library "pygames"</b>
-The Game is based on pygames. To run it you need to install the pygames library.
-You can do so by typing "pip install pygame" in a terminal within your IDE.
-<h2>Manual</h2>
-There is not much to learn at the moment.
-Just use the ARROW KEYS of your keyboard to fly around the screen and try to survice as long as possible.
-To exit the game press ESC.
+```
+
+Since this game is based an "pygames" you need to install the pygames library. To do so open a terminal within your IDE and type
+
+```sh
+pip install pygame
+```
+
+If you are cloning the project then run this first, otherwise you can download the source code and skip this step.
+
+```sh
+git clone https://github.com/pablee06/pablo_github.git
+```
+
+## Usage ℹ️
+
+To start playing go to the root of your folder and run this command
+
+```sh
+python spacegame.ty
+```
+
+Within the game just use the 'ARROW KEYS' of your keyboard to fly around the screen.
+Try to survice as long as possible to crack the highscore.
+
+To exit the game press 'ESC' on your kyboard.
+
 Have fun!
